Step4 - 로또(수동) (#2578)
* refactor: step3 리뷰 반영; 생성자 위치 조정

* refactor: step3 리뷰 반영; 보너스 로또 번호만 수용할 수 있게 수정

* refactor: step3 리뷰 반영; ticket에서 result 객체 참조 하도록 수정

* feat: 수동 로또 구매 수 입력

* feat: 수동으로 로또 숫자 입력

* feat: 자동 & 수동 티켓 갯수 출력

* refactor: 티켓 구입 검증 위치 변경

* bugfix: Money 객체 자료형 수정

- 대량의 입력이 들어올 경우, BigInteger 객체로 대체 필요

* refactor: 코드 리포맷팅

* refactor: User 생성 및 검증 수행 위치 이동

* refactor: 수동 로또 숫자 입력 위치 변경

- TODO: TicketSeller에 수동 로또 로직 반영

* refactor: 수동 로또 티켓 생성 로직 변경

* refactor: User 검증 위치 이동 및 테스트 보강

+++ src/main/java/lottogame/domain/Money.java
@@ -6,9 +6,9 @@ public class Money {
     private static final String NEGATIVE_AMOUNT_EXCEPTION_MESSAGE = "돈의 액수는 음수가 될 수 없습니다.";
     public static final Money ZERO = new Money(0);
 
-    private final int money;
+    private final long money;
 
-    public Money(int money) {
+    public Money(long money) {
         validate(money);
         this.money = money;
     }
@@ -29,13 +29,13 @@ public class Money {
         return money >= other.money;
     }
 
-    private void validate(int money) {
+    private void validate(long money) {
         if (money < 0) {
             throw new IllegalArgumentException(NEGATIVE_AMOUNT_EXCEPTION_MESSAGE);
         }
     }
 
-    public int getValue() {
+    public long getValue() {
         return money;
     }
